update:add hotel card on home page with backend data

+++ backend/controllers/hotel-details-controller.js
@@ -1,5 +1,5 @@
 import z from 'zod'
-import { uploadImageOnCloudinary } from '../utils/cloudinary.js'
+import { deleteImageOnCloudinary, uploadImageOnCloudinary } from '../utils/cloudinary.js'
 import { hotelDetailsModel } from '../model/hotel-details-model.js'
 
 const hotelDetailsValidation=z.object({
@@ -21,12 +21,12 @@ const hotelDetailsValidation=z.object({
     bookingType:z.string({
         required_error:"BookingType is required"
     }).trim(),
-    price:z.string({
+    price:z.number({
         required_error:"Price is required"
-    }).trim(),
-    discount:z.string({
+    }),
+    discount:z.number({
         required_error:"Discount is required"
-    }).trim(),
+    }),
     customerNumber:z.object({
         bed:z.number(),
         guest:z.number(),
@@ -47,29 +47,29 @@ export const hotelDetailsController=async(req,resp)=>{
     let hotelImage=req.files
 
     offerServices=JSON.parse(offerServices)
+    console.log(offerServices);
     customerNumber=JSON.parse(customerNumber)
     locatedPlace=JSON.parse(locatedPlace)
-    discount=JSON.parse(discount)
+    discount=JSON.parse(JSON.parse(discount))
+    price=JSON.parse(JSON.parse(price))
     roomType=JSON.parse(roomType)
     houseTitle=JSON.parse(houseTitle)
-    price=JSON.parse(price)
     const validateDetails=hotelDetailsValidation.safeParse({homeType,roomType,houseTitle,aboutHome,offerServices,description,customerNumber,bookingType,price,discount,locatedPlace})
    
     if(validateDetails.error){
         return resp.json({error:validateDetails.error.format()})
     }
 
-    const uploadImage=await uploadImageOnCloudinary(hotelImage,'hotelDetails')
-    console.log('see:  ',uploadImage);
+    const resultOfUploadedImage=await uploadImageOnCloudinary(hotelImage,'hotelDetails')
     
-    if(!uploadImage || uploadImage.length === 0){
+    if(!resultOfUploadedImage || resultOfUploadedImage.length === 0){
         throw new Error("Image not uploaded")
     }
-    let hotelImagesUrl=uploadImage.map((item)=>item.url)
-    console.log(hotelImagesUrl);
+    // const idOfImage=resultOfUploadedImage.map(({url,public_id})=>({url,public_id}))
+    const idOfImage=resultOfUploadedImage.map((item)=>({url:item.url,public_id:item.public_id}))
     const uploadedBy=req.user._id
     
-    const hotelDataOnDb=await hotelDetailsModel.create({uploadedBy,hotelImagesUrl,homeType,roomType,houseTitle,aboutHome,offerServices,description,customerNumber,bookingType,price,discount,locatedPlace})
+    const hotelDataOnDb=await hotelDetailsModel.create({uploadedBy,idOfImage,homeType,roomType,houseTitle,aboutHome,offerServices,description,customerNumber,bookingType,price,discount,locatedPlace})
     if(!hotelDataOnDb){
         return resp.json({success:false,message:"Details are not store in db"})
     }
@@ -91,7 +91,6 @@ export const getHotelDetailsController=async(req,resp)=>{
             return resp.json({success:false,message:"There is no hotel register by this user"})
         }
 
-        console.log(hotelDetails);
         return resp.json({success:true,detals:hotelDetails})
     } catch (error) {
         resp.json({success:false,message:"Error: "+error.message})
@@ -99,22 +98,50 @@ export const getHotelDetailsController=async(req,resp)=>{
 }
 
 export const deleteHotelController=async(req,resp)=>{
-    const id=req.params.id
-    console.log('id: ',id);
+    let {id}=req.body
     try {
         if(!id){
             return resp.json({success:false,message:"Please provide id"})
         }
+        const findDetails=await hotelDetailsModel.find({_id:{$in:id}})
     
+        const deleteImage=await deleteImageOnCloudinary(findDetails)
+        if(!deleteImage){
+            return resp.json({success:false,message:"Image is not deleted from cloudinary"})
+        }
         const result=await hotelDetailsModel.deleteMany({_id:{$in:id}})
+
+        if(result.deletedCount===0){
+            return resp.json({success:false,message:"Details not deleted"})
+        }
+        
 
         if(!result){
             return resp.json({success:false,message:"Details not found"})
         }
 
-        console.log('deleted');
         return resp.json({success:true,message:"Deleted Successfully",details:result})
     } catch (error) {
         return resp.json({success:false,message:"Error: "+error.message})
     }
+}
+
+
+export const getSingleDetails=async(req,resp)=>{
+  try {
+    const{id}=req.params
+  if(!id){
+    return resp.json({success:false,message:"Please provide id first"})
+  }
+
+  const result=await hotelDetailsModel.find({_id:id})
+  console.log(result);
+  if(!result){
+    resp.json({success:false,message:"Details not found"})
+  }
+
+  return resp.json({success:true,message:"Detail find Successfully",data:result})
+  } catch (error) {
+    resp.json({success:false,message:"Error: "+error.message})
+  }
 }
